Lightroom: optimized CC rendering, fixed displayed LUT colors | MXAO: Minimal UI improvement
The filter now renders all its CC to a 4096x64 LUT and reads that afterwards, as opposed to processing every pixel on the screen. This results in only 12.6% of the workload for color correction ops only on a full HD resolution - workload is now independent from screen resolution.

Fixed the displayed LUT colors, apparently on DX11, the colors came out differently as they did on DX9. There _was_ a bug in the code but had no effect on DX9 so I assumed the code to be working fine.

Changed MXAO blend type drag type to slider, looks cleaner that way.